modified:   TODOS/main.py 	new file:   TODOS/routers/admin.py 	modified:   TODOS/routers/auth.py 	modified:   TODOS/routers/todos.py 	modified:   TODOS/todosapp.db

+++ TODOS/routers/todos.py
@@ -33,6 +33,7 @@ async def read_all(user: user_dependancy, db: Annotated[Session, Depends(get_db)
 
 @router.get('/todo/{todo_id}')
 async def read_by_id(user: user_dependancy, db: db_dependancy, todo_id: int = Path(gt=0)):
+    
     if user is None:
         raise HTTPException(status_code=401, detail='Authentication Failed')
     
@@ -53,10 +54,15 @@ async def create_todo(user: user_dependancy, db: db_dependancy, todo_request: To
     db.commit()
     
 @router.put('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependancy, todo_request: TodoRequest,  todo_id: int = Path(gt=0)):
+async def update_todo(user: user_dependancy, db: db_dependancy, todo_request: TodoRequest,  todo_id: int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    
     todo_model = db.query(ToDos).filter(ToDos.id == todo_id).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
+
     todo_model.title = todo_request.title # type: ignore
     todo_model.description = todo_request.description # type: ignore
     todo_model.priority = todo_request.priority # type: ignore
@@ -66,9 +72,14 @@ async def update_todo(db: db_dependancy, todo_request: TodoRequest,  todo_id: in
     db.commit()
     
 @router.delete('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependancy, todo_id: int = Path(gt=0)):
-    todo_model = db.query(ToDos).filter(ToDos.id == todo_id).first()
+async def delete_todo(user: user_dependancy, db: db_dependancy, todo_id: int = Path(gt=0)):
+    
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')    
+    
+    todo_model = db.query(ToDos).filter(ToDos.id == todo_id).filter(ToDos.owner_id == user.get('id')).first()
+    
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
-    db.query(ToDos).filter(ToDos.id == todo_id).delete()
+    db.query(ToDos).filter(ToDos.id == todo_id).filter(ToDos.owner_id == user.get('id')).delete()
     db.commit()
